refactor: 구분자 처리 로직을 별도 클래스로 분리
README.md 구현에 맞춰 최신화
