+++ README.md
@@ -8,7 +8,8 @@
         - 커스텀 구분자를 지정할 수 있다.
             - 커스텀 구분자는 문자열 앞부분의 "//"와 "\n" 사이에 위치한다.
 - [x] 계산기
-    - [x] 문자열을 기본 구분자, 커스텀 구분자로 분리한다.
+    - [x] 구분자 처리기
+        - [x] 문자열을 기본 구분자, 커스텀 구분자로 분리한다.
     - [x] 분리된 문자열을 정수로 변환한다.
     - [x] 정수로 변환한 값을 더한다.
     - [x] 더한 결과를 반환한다.
